Restrict actions if not logged in

+++ src/components/MarkWatchedMovie.js
@@ -13,6 +13,8 @@ function MarkWatchedMovie({isWatched, movieId}) {
     }, [isWatched]);
 
     function handleWatchedButton() {
+        if (!sessionStorage.getItem(GlobalStorageKeys.USER_IS_LOGGED_IN)) return;
+        
         if (isWatched) {
             fetch(`${process.env.REACT_APP_API_URL}/watchedMovies/${movieId}`, {
                 method: 'DELETE',
